Fix truthy tests for numbers
Values with absolute value < 1 weren't being treated as truthy because
of accidental rounding to zero.

+++ src/main/java/org/haplo/template/driver/jrubyjson/JRubyJSONDriver.java
@@ -103,7 +103,7 @@ class JRubyJSONDriver extends Driver {
             } else if(o instanceof RubyArray) {
                 return !(((RubyArray)o).isEmpty());
             } else if(value instanceof RubyNumeric) {
-                return ((RubyNumeric)value).getLongValue() != 0;
+                return ((RubyNumeric)value).getDoubleValue() != 0.0;
             }
             return false;
         } else {
